pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "life-is-beautiful-podcast"
version = "0.1.0"
description = "Generate podcasts from Life is Beautiful newsletter markdown content"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Tomodo1773", email = "pomo1773@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-genai>=0.1.0",
    "pydub>=0.25.1",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "tests"]

[tool.ruff]
line-length = 127
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "C4", "SIM", "RUF"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"app/utils/podcast_generator.py" = ["RUF001", "E501", "W291"]
